Car Color Detection Model/requirements.txt
numpy
pandas
opencv-python
Pillow
tensorflow
streamlit
ultralytics
matplotlib
seaborn
scikit-learn
